Xpath extension (#506)
* Add book and page xpath search view

Add `/xpath?id=<book-or-page-uuid>&q=<xpath-query>` for doing xpath search
on index.cnxml.

* Change xpath to respond to .html or .json

Instead of checking whether the html request "Accept" header, just use
the path to decide whether to return html or json.

* Refactor execute_xpath in views/xpath.py

Return result from database instead of doing all the request and
response stuff.

* Return the json result in html format for xpath.html

xpath.html was returning the entire book tree whether the page contains
the xpath or not.  xpath.json uses a completely different sql for
getting the results.  Change them to use the same sql.

* Find nearest ancestor with id in xpath search

When user uses the xpath search, add `/ancestor-or-self::*[@id][1]` to
the end of their query to find nearest ancestor with an id so we can
generate a link to approximately that part of the document.

Change list link text to text without xml, and if there is no text,
display the tag.

* Fetch individual index.cnxml for book xpath queries

Instead of just doing xpath in the sql query, fetch the index.cnxml for
each page that the sql query returns, and do the xpath search in python
so that we can get the closest ancestor id.

I used this for benchmarking:

```
time wget -O /dev/null 'http://localhost:6500/xpath.html?q=//*[name()=%27link%27]&id=185cbf87-c72e-48f5-b51e-f14f21b5eabd@10.115'
```

The time it took in the previous version was 0.891 seconds, and this
version took 2.961 seconds.  It is about 3.3 times slower.

* Add namespaces to xpath search

This allows xpath search queries to use paths like `//cnx:definition`
instead of `//*[name()='definition']`.