Use cnx-db docker image in travis tests (#521)
* Add docker files for archive

Build a docker image for archive which works with cnx-db and memcached.

* Use docker images on travis

On travis trusty, the default Linux environment, postgres is probably
running in a different server.  It was not able to see the files in
/home/travis/virtualenv/python2.7.13, which is where our source code is
checked out on travis.

* Skip test_pyimport if database is not on localhost

test_pyimport checks a module file path returned by a function in the
database is the same as the one installed locally.  This only works if
the database and archive are running on the same host.  Add a condition
to skip this if database is not on localhost.