Fix number of migrations to rollback in ci_test_migrations.sh
When calculating the number of migrations to rollback,
ci_test_migrations.sh does not take into account the repeat, deferred
migrations that ran again, so dbmigrator was not rolling back enough
migrations for the rollback test.

In this particular case, we have one normal migration
`20171005213556_first-last` and one repeat, deferred migration
`20171010195430_install-recipes`.  dbmigrator was only rolling back the
deferred migration and not the normal migration.

Changing the script to mark repeat, deferred migrations as not completed
should fix the calculation of the number of migrations.