Use plpython3u for python 3 tests on travis
There are strange errors if we use a python 3 virtualenv together with
python 2 on postgresql.  The first thing that fails is when initializing
venv on postgresql, we try to add `.../lib/python2.7/site-packages` to
`sys.path` which is of course incorrect in a python3.5 virtualenv.

Use `sed` to change all the `plpythonu` to `plpython3u` before running
the tests.