Use virtualenv python version in venv site packages path
When activating venv in postgres, one of the steps is to add the
virtualenv site packages path to `sys.path`, and the path uses the
python version which might be different between the virtualenv python
and the postgres python.  The correct site packages path is to use the
virtualenv python version.