Use our build of the postgres container image
This uses our build of the postgres:9.4 container because the existing release of the container is not up-to-date. This causes issues with our `RUN set -x && apt-get update && ...` which then tries to upgrade postgres, but is apparently failing. This is then causing the container to refuse external connections to postgres.