Change `CREATE ROLE` to `CREATE USER` in initdb.sh
The `$DB_USER` postgres user we created is not permitted to log in.

From the postgres `CREATE ROLE` documentation
https://www.postgresql.org/docs/9.4/static/sql-createrole.html

About `NOLOGIN` / `LOGIN`:

These clauses determine whether a role is allowed to log in; that is,
whether the role can be given as the initial session authorization name
during client connection. A role having the LOGIN attribute can be
thought of as a user. Roles without this attribute are useful for
managing database privileges, but are not users in the usual sense of
the word. If not specified, NOLOGIN is the default, except when CREATE
ROLE is invoked through its alternative spelling CREATE USER.

So we will just switch to using `CREATE USER` instead of having to
specify `LOGIN` with `CREATE ROLE`.