Fix venv importable tests to skip if told
This is being fixed because one might be in a venv, but not want to connect to the database as virtualenv importable.

This could not make sense, but for my usecase it works. I'm within a venv but running the tests through a container mounted with the local code. The container is not setup to work with a venv, because it doesn't need to be. This change simply brings both scenarios into a compatible use via the `AS_VENV_IMPORTABLE` environment variable.