Add a read-only database setting
This is to be used for operations that only required read access from the database. This means we can configure the application to use the write connections ('common') only when necessary.

We already have a setup that utilizes the replicants for known read-only operations. This now enables the developer to use the correct connection type for context.