Fix settings discover to use given settings if necessary
This fixes an issue when a set of settings is given and the env vars are not set. This ensure that we use the given value for the `db.common.url` setting in place of the missing DB_URL rather than leave it `None`, which would set the defaults for the other urls to `None` as well.