Move the test condition func into a common module
This moves the `py3_too_old` function into the testing module.

This also re-enables the test the `skipif` was decorating, because the `skipif` will consume a callable for its arguments, but this wasn't the indended usage for the defined function. So this instead calls the function and thus enables the test.