Recreate connection pool after init_venv in tests
We use sqlalchemy to initialize our database, that includes setting a
function that is run at the beginning of every session (what init_venv
does).  But sqlalchemy engines have connection pools that are already
connected, so virtualenv is not set up correctly for those connections,
causing the triggers to not be able to find the cnx code.