Update ident hash code to be py3 compatible
Most of the changes have to do with python 3 treating bytes and str
differently from python 2 so now you have to do str operations with
str and cannot mix with bytes.

In python 2, the uuid class had `get_hex()` and `get_bytes()` which are
the same as `.hex` and `.bytes`, in python 3, the `get_` methods are
removed, so change the code to use `.hex` and `.bytes`.