Remove transforms code (moved to cnx-transforms) (#176)
Since `neb cnxml-to-html` needs to use the transforms code, and we
decide that we don't want to have cnx-db as a dependency, we are
splitting out the transforms code into a separate package called
cnx-transforms.  This means we won't have any database code when all we
want is to transform cnxml to html.

This also adds a migration to change the imports in triggers to point to
cnx-transforms.