Separate linting from package installation
This separates the linting setup and usage from the package's setup and usage.
When developing on this project within a container, one doesn't have the
necessary system requirements to build `psycopg2` (and `psycopg2-binary` can't
be used because the `psycopg2` dependency is mentioned upstream and
downstream). To correct this issue, this change enables the developer to
lint the package locally without actually installing the package.