Repeat populate tree slug migration
We adjusted our implementation of generating slugs in cnx-common so we
need to regenerate all the slugs in the database.

Adding `should_run` in `populate_trees_slug_values` migration means that
dbmigrator is going to check it to see if this migration should run.
The condition is based on the bug report we received: if the slug for
`Astronomy` -> `2 Observing the Sky: The Birth of Astronomy` ->
`Exercises` -> `Review Questions` is `review-questions` (instead of
`2-review-questions`), the migration should run.

We can remove the `should_run` from this migration once this has been
deployed to prod.