Fix empty collated file associations in dump_book.py (#185)
I assumed that the first module_ident returned in
`get_module_idents_tree_nodes` is the book module ident, which I then
use to download `collated_file_associations`.  This is an incorrect
assumption probably due to `SELECT DISTINCT` in the sql in
`get_module_idents_tree_nodes`.  Because of this,
`collated_file_assocations` is empty.

To fix this, change `get_module_idents_tree_nodes` to order by `parent`
descending, which should get the book, with `parent` null, to the top.