Rebake books in cnxml-to-html transforms migration (#187)
Whenever rhaptos.cnxmlutils is updated, transforms is run automatically.
But only the raw html is updated, the baked html is untouched.  The raw
html isn't what we serve for most of the openstax books, so we have to
rebake the books to make sure the changes also get to the baked html.