Separate db_tables function and fixture for reuse
In cnx-press, the `db_tables` fixture is used as a session fixture
instead of a function fixture but pytest 4.1.0 make it an error to call
a fixture:

```
Fixture "db_tables" called directly. Fixtures are not meant to be called directly,
but are created automatically when test functions request them as parameters.
See https://docs.pytest.org/en/latest/fixture.html for more information about fixtures, and
https://docs.pytest.org/en/latest/deprecations.html#calling-fixtures-directly about how to update your code.
```

One of the solutions suggested by pytest is to separate the function that has
the fixture decorator and the function that returns the fixture.

This fixes cnx-press so it can use the latest version of pytest.