Removed 20170911201035_transform_cnxml_to_html.py from migrations
Based on a situation we had back in [Nov/Dec 2019](https://github.com/openstax/cnx/issues/766#issuecomment-564142344) it was found that when
cnxmlutils was updated this deferred migration would run and transform
the html of all the content. This caused problems for Unified because of the assumption
that our content was immutable. We've removed this migration now b/c we
are less concerned with updating old versions, rhaptos.cnxmlutils is
stable, new versions will continue to get the new transforms applied,
and archive will eventually be rewritten.